requirements.txt
streamlit
tensorflow
pandas
numpy
plotly
scikit-learn
Pillow
